+++ packages/manager/src/mocks/presets/crud/handlers/vpcs.ts
@@ -120,6 +120,7 @@ export const createVPC = (mockState: MockState) => [
             ...subnetPayload,
             created: DateTime.now().toISO(),
             linodes: [],
+            nodebalancers: [],
             updated: DateTime.now().toISO(),
           });
           vpcSubnets.push(subnet);
