+++ packages/manager/src/mocks/types.ts
@@ -6,6 +6,7 @@ import type {
   Event,
   Firewall,
   FirewallDevice,
+  Interface,
   IPAddress,
   KubeNodePoolResponse,
   KubernetesCluster,
@@ -155,16 +156,17 @@ export type MockHandler = (mockState: MockState) => HttpHandler[];
  */
 export interface MockState {
   cloudnats: CloudNAT[];
+  configInterfaces: [number, Interface][]; // number is Config ID
   domainRecords: DomainRecord[];
   domains: Domain[];
   eventQueue: Event[];
-  firewallDevices: [number, FirewallDevice][];
+  firewallDevices: [number, FirewallDevice][]; // number is Firewall ID
   firewalls: Firewall[];
   ipAddresses: IPAddress[];
   kubernetesClusters: KubernetesCluster[];
   kubernetesNodePools: KubeNodePoolResponse[];
-  linodeConfigs: [number, Config][];
-  linodeInterfaces: [number, LinodeInterface][];
+  linodeConfigs: [number, Config][]; // number is Linode ID
+  linodeInterfaces: [number, LinodeInterface][]; // number is Linode ID
   linodes: Linode[];
   nodeBalancerConfigNodes: NodeBalancerConfigNode[];
   nodeBalancerConfigs: NodeBalancerConfig[];
@@ -173,7 +175,7 @@ export interface MockState {
   placementGroups: PlacementGroup[];
   regionAvailability: RegionAvailability[];
   regions: Region[];
-  subnets: [number, Subnet][];
+  subnets: [number, Subnet][]; // number is VPC ID
   supportReplies: SupportReply[];
   supportTickets: SupportTicket[];
   volumes: Volume[];
